requirements.txt
requests>=2.32.0
pydantic>=2.6.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
